+++ sdk/trace/provider_test.go
@@ -30,9 +30,9 @@ func (t *basicSpanProcesor) Shutdown(context.Context) error {
 	return nil
 }
 
-func (t *basicSpanProcesor) OnStart(s *export.SpanData) {}
-func (t *basicSpanProcesor) OnEnd(s *export.SpanData)   {}
-func (t *basicSpanProcesor) ForceFlush()                {}
+func (t *basicSpanProcesor) OnStart(parent context.Context, s *export.SpanData) {}
+func (t *basicSpanProcesor) OnEnd(s *export.SpanData)                           {}
+func (t *basicSpanProcesor) ForceFlush()                                        {}
 
 func TestShutdownTraceProvider(t *testing.T) {
 	stp := NewTracerProvider()